upload all

+++ demo (12)/demo/src/main/java/com/example/demo/service/ClassRoomService.java
@@ -36,24 +36,30 @@ public class ClassRoomService implements IClassRoomService{
 
     @Override
     public void updateClassRoom(ClassRoom classRoom, Integer id) throws ClientFaultException, JsonProcessingException {
-        if (classRoom.getClassRoomType().equals(ClassRoomType.EXTERNAL) && classRoom.getNumberOfStudents() >= maxstudents){
-            throw new NotIllegalChangeClassType("Cannot change the class room type because there are more than 20 students");
-        }
-        classRoomRepository.updateClassRoom(classRoom, id);
-        if (cache_on && cacheRepository.isKeyExist("CLASS" + id.toString())) {
-            cacheRepository.updateCacheEntity("CLASS" + id.toString(), objectMapper.writeValueAsString(classRoom));
+        ClassRoom classRoomUpdate = classRoomRepository.getClassRoomById(id);
+        if (classRoomUpdate != null) {
+            if (classRoom.getClassRoomType().equals(ClassRoomType.EXTERNAL) && classRoom.getNumberOfStudents() >= maxstudents) {
+                throw new NotIllegalChangeClassType("Cannot change the class room type because there are more than 20 students");
+            }
+            classRoomRepository.updateClassRoom(classRoom, id);
+            if (cache_on && cacheRepository.isKeyExist("CLASS" + id.toString())) {
+                cacheRepository.updateCacheEntity("CLASS" + id.toString(), objectMapper.writeValueAsString(classRoom));
+            }
         }
     }
 
 
     @Override
     public void deleteClassRoom(Integer id) throws ClientFaultException {
-        if(classRoomRepository.getClassRoomById(id).getNumberOfStudents() >= 1 ){
-            throw new ClassNotEmptyException("Cannot delete this classRoom because its not empty ");
-        }
-        classRoomRepository.deleteClassRoom(id);
-        if (cache_on && cacheRepository.isKeyExist("CLASS" + id.toString())) {
-            cacheRepository.removeKey("CLASS" + id.toString());
+        ClassRoom classRoomDelete = classRoomRepository.getClassRoomById(id);
+        if (classRoomDelete != null) {
+            if (classRoomRepository.getClassRoomById(id).getNumberOfStudents() >= 1) {
+                throw new ClassNotEmptyException("Cannot delete this classRoom because its not empty ");
+            }
+            classRoomRepository.deleteClassRoom(id);
+            if (cache_on && cacheRepository.isKeyExist("CLASS" + id.toString())) {
+                cacheRepository.removeKey("CLASS" + id.toString());
+            }
         }
     }
 
@@ -74,9 +80,10 @@ public class ClassRoomService implements IClassRoomService{
             }
 
             ClassRoom result = classRoomRepository.getClassRoomById(id);
-
-            if (cache_on) {
-                cacheRepository.createCacheEntity("CLASS" + id.toString(), objectMapper.writeValueAsString(result));
+            if(result != null) {
+                if (cache_on) {
+                    cacheRepository.createCacheEntity("CLASS" + id.toString(), objectMapper.writeValueAsString(result));
+                }
             }
             return result;
